Address PR feedback from @cleemullins
