+++ libraries/Microsoft.Bot.Builder.Core.State/ConversationStateManager.cs
@@ -10,8 +10,6 @@ namespace Microsoft.Bot.Builder.Core.State
 
     public class ConversationStateManager : StateManager, IConversationStateManager
     {
-        public static string PreferredStateStoreName = $"{nameof(ConversationStateManager)}.PreferredStateStore";
-
         public ConversationStateManager(string channelId, string conversationId, IStateStorageProvider stateStore) : base(BuildStateNamespace(channelId, conversationId), stateStore)
         {
             ChannelId = channelId ?? throw new System.ArgumentNullException(nameof(channelId));
@@ -23,8 +21,29 @@ namespace Microsoft.Bot.Builder.Core.State
 
         public string ConversationId { get; }
 
-        public static string BuildStateNamespace(string channelId, string conversationId) => $"/channels/{channelId}/conversations/{conversationId}";
+        public static string BuildStateNamespace(string channelId, string conversationId)
+        {
+            if (string.IsNullOrEmpty(channelId))
+            {
+                throw new ArgumentException("Expected a non-null/empty value.", nameof(channelId));
+            }
 
-        public static string BuildStateNamespace(ITurnContext turnContext) => BuildStateNamespace(turnContext.Activity.ChannelId, turnContext.Activity.Conversation.Id);
+            if (string.IsNullOrEmpty(conversationId))
+            {
+                throw new ArgumentException("Expected a non-null/empty value.", nameof(conversationId));
+            }
+
+            return $"/channels/{channelId}/conversations/{conversationId}";
+        }
+
+        public static string BuildStateNamespace(ITurnContext turnContext)
+        {
+            if (turnContext == null)
+            {
+                throw new ArgumentNullException(nameof(turnContext));
+            }
+
+            return BuildStateNamespace(turnContext.Activity.ChannelId, turnContext.Activity.Conversation.Id);
+        }
     }
 }
