+++ libraries/Microsoft.Bot.Builder.Core.State/IStateManagerServiceResolver.cs
@@ -1,0 +1,10 @@
+﻿namespace Microsoft.Bot.Builder.Core.State
+{
+    public interface IStateManagerServiceResolver
+    {
+        TStateManager ResolveStateManager<TStateManager>() where TStateManager : class, IStateManager;
+        IStateManager ResolveStateManager(string stateNamespace);
+    }
+
+
+}
